Refactor type resolver trait into functions.
