+++ sway-core/src/semantic_analysis/type_check_context.rs
@@ -6,10 +6,10 @@ use crate::{
     engine_threading::*,
     language::{
         parsed::TreeType,
-        ty::{self, TyDecl, TyTraitItem},
+        ty::{self, TyDecl},
         CallPath, QualifiedCallPath, Visibility,
     },
-    monomorphization::{monomorphize_with_modpath, type_decl_opt_to_type_id, MonomorphizeHelper},
+    monomorphization::{monomorphize_with_modpath, MonomorphizeHelper},
     namespace::{
         IsExtendingExistingImpl, IsImplSelf, ModulePath, ResolvedDeclaration,
         ResolvedTraitImplItem, TraitMap, TryInsertingTraitImplOnFailure,
@@ -26,10 +26,10 @@ use sway_error::{
     handler::{ErrorEmitted, Handler},
 };
 use sway_types::{span::Span, Ident, Spanned};
-use sway_utils::iter_prefixes;
 
 use super::{
-    symbol_collection_context::SymbolCollectionContext, type_resolve::TypeResolver,
+    symbol_collection_context::SymbolCollectionContext,
+    type_resolve::{resolve, resolve_call_path, resolve_qualified_call_path},
     GenericShadowingMode,
 };
 
@@ -584,13 +584,14 @@ impl<'a> TypeCheckContext<'a> {
             handler,
             self.engines(),
             self.namespace(),
-            self,
             value,
             type_arguments,
             enforce_type_arguments,
             call_site_span,
             &mod_path,
             self.self_type(),
+            &self.type_subst(),
+            &SubstTypesContext::new(self.engines(), !self.collecting_unifications()),
         )
     }
 
@@ -656,20 +657,18 @@ impl<'a> TypeCheckContext<'a> {
         enforce_type_arguments: EnforceTypeArguments,
         type_info_prefix: Option<&ModulePath>,
     ) -> Result<TypeId, ErrorEmitted> {
-        let engines = self.engines();
-        let namespace = self.namespace();
-        let mod_path = self.namespace().mod_path.clone();
-        let self_type = self.self_type();
-        self.resolve(
+        resolve(
             handler,
-            engines,
-            namespace,
-            &mod_path,
+            self.engines(),
+            self.namespace(),
+            &self.namespace().mod_path,
             type_id,
             span,
             enforce_type_arguments,
             type_info_prefix,
-            self_type,
+            self.self_type(),
+            &self.type_subst(),
+            &SubstTypesContext::new(self.engines(), !self.collecting_unifications()),
         )
     }
 
@@ -679,12 +678,13 @@ impl<'a> TypeCheckContext<'a> {
         handler: &Handler,
         call_path: &CallPath,
     ) -> Result<ty::TyDecl, ErrorEmitted> {
-        self.resolve_call_path(
+        resolve_call_path(
             handler,
             self.engines(),
             self.namespace(),
             &self.namespace().mod_path,
             call_path,
+            self.self_type(),
         )
     }
 
@@ -693,12 +693,15 @@ impl<'a> TypeCheckContext<'a> {
         handler: &Handler,
         qualified_call_path: &QualifiedCallPath,
     ) -> Result<ty::TyDecl, ErrorEmitted> {
-        self.resolve_qualified_call_path(
+        resolve_qualified_call_path(
             handler,
             self.engines(),
             self.namespace(),
             &self.namespace().mod_path.clone(),
             qualified_call_path,
+            self.self_type(),
+            &self.type_subst(),
+            &SubstTypesContext::new(self.engines(), !self.collecting_unifications()),
         )
     }
 
@@ -734,21 +737,20 @@ impl<'a> TypeCheckContext<'a> {
             .get_items_for_type(self.engines, type_id);
 
         // resolve the type
-        let type_id = self
-            .resolve(
-                handler,
-                self.engines(),
-                self.namespace(),
-                item_prefix,
-                type_id,
-                &item_name.span(),
-                EnforceTypeArguments::No,
-                None,
-                self.self_type(),
-            )
-            .unwrap_or_else(|err| {
-                type_engine.insert(self.engines, TypeInfo::ErrorRecovery(err), None)
-            });
+        let type_id = resolve(
+            handler,
+            self.engines(),
+            self.namespace(),
+            item_prefix,
+            type_id,
+            &item_name.span(),
+            EnforceTypeArguments::No,
+            None,
+            self.self_type(),
+            &self.type_subst(),
+            &SubstTypesContext::new(self.engines(), !self.collecting_unifications()),
+        )
+        .unwrap_or_else(|err| type_engine.insert(self.engines, TypeInfo::ErrorRecovery(err), None));
 
         // grab the module where the type itself is declared
         let type_module = self.namespace().lookup_submodule_from_absolute_path(
@@ -1286,347 +1288,5 @@ impl<'a> TypeCheckContext<'a> {
                 code_block_first_pass.into(),
             )
             .is_ok()
-    }
-}
-
-impl TypeResolver for TypeCheckContext<'_> {
-    /// Resolve the type of the given [TypeId], replacing any instances of
-    /// [TypeInfo::Custom] with either a monomorphized struct, monomorphized
-    /// enum, or a reference to a type parameter.
-    #[allow(clippy::too_many_arguments)]
-    fn resolve(
-        &self,
-        handler: &Handler,
-        engines: &Engines,
-        namespace: &Namespace,
-        mod_path: &ModulePath,
-        type_id: TypeId,
-        span: &Span,
-        enforce_type_arguments: EnforceTypeArguments,
-        type_info_prefix: Option<&ModulePath>,
-        self_type: Option<TypeId>,
-    ) -> Result<TypeId, ErrorEmitted> {
-        let type_engine = engines.te();
-        let module_path = type_info_prefix.unwrap_or(mod_path);
-        let type_id = match (*type_engine.get(type_id)).clone() {
-            TypeInfo::Custom {
-                qualified_call_path,
-                type_arguments,
-                root_type_id,
-            } => {
-                let type_decl_opt = if let Some(root_type_id) = root_type_id {
-                    namespace
-                        .root()
-                        .resolve_call_path_and_root_type_id(
-                            handler,
-                            engines,
-                            namespace.module(engines),
-                            root_type_id,
-                            None,
-                            &qualified_call_path.clone().to_call_path(handler)?,
-                            self.self_type(),
-                        )
-                        .map(|decl| decl.expect_typed())
-                        .ok()
-                } else {
-                    self.resolve_qualified_call_path(
-                        handler,
-                        engines,
-                        namespace,
-                        module_path,
-                        &qualified_call_path,
-                    )
-                    .ok()
-                };
-                type_decl_opt_to_type_id(
-                    handler,
-                    engines,
-                    namespace,
-                    self,
-                    type_decl_opt,
-                    &qualified_call_path.call_path,
-                    span,
-                    enforce_type_arguments,
-                    mod_path,
-                    type_arguments.clone(),
-                    self_type,
-                )?
-            }
-            TypeInfo::Array(mut elem_ty, n) => {
-                elem_ty.type_id = self
-                    .resolve(
-                        handler,
-                        engines,
-                        namespace,
-                        mod_path,
-                        elem_ty.type_id,
-                        span,
-                        enforce_type_arguments,
-                        None,
-                        self_type,
-                    )
-                    .unwrap_or_else(|err| {
-                        engines
-                            .te()
-                            .insert(engines, TypeInfo::ErrorRecovery(err), None)
-                    });
-
-                engines.te().insert(
-                    engines,
-                    TypeInfo::Array(elem_ty.clone(), n.clone()),
-                    elem_ty.span.source_id(),
-                )
-            }
-            TypeInfo::Slice(mut elem_ty) => {
-                elem_ty.type_id = self
-                    .resolve(
-                        handler,
-                        engines,
-                        namespace,
-                        mod_path,
-                        elem_ty.type_id,
-                        span,
-                        enforce_type_arguments,
-                        None,
-                        self_type,
-                    )
-                    .unwrap_or_else(|err| {
-                        engines
-                            .te()
-                            .insert(engines, TypeInfo::ErrorRecovery(err), None)
-                    });
-
-                engines.te().insert(
-                    engines,
-                    TypeInfo::Slice(elem_ty.clone()),
-                    elem_ty.span.source_id(),
-                )
-            }
-            TypeInfo::Tuple(mut type_arguments) => {
-                for type_argument in type_arguments.iter_mut() {
-                    type_argument.type_id = self
-                        .resolve(
-                            handler,
-                            engines,
-                            namespace,
-                            mod_path,
-                            type_argument.type_id,
-                            span,
-                            enforce_type_arguments,
-                            None,
-                            self_type,
-                        )
-                        .unwrap_or_else(|err| {
-                            engines
-                                .te()
-                                .insert(engines, TypeInfo::ErrorRecovery(err), None)
-                        });
-                }
-
-                engines
-                    .te()
-                    .insert(engines, TypeInfo::Tuple(type_arguments), span.source_id())
-            }
-            TypeInfo::TraitType {
-                name,
-                trait_type_id,
-            } => {
-                let item_ref = self.namespace().get_root_trait_item_for_type(
-                    handler,
-                    engines,
-                    &name,
-                    trait_type_id,
-                    None,
-                )?;
-                if let ResolvedTraitImplItem::Typed(TyTraitItem::Type(type_ref)) = item_ref {
-                    let type_decl = engines.de().get_type(type_ref.id());
-                    if let Some(ty) = &type_decl.ty {
-                        ty.type_id
-                    } else {
-                        type_id
-                    }
-                } else {
-                    return Err(handler.emit_err(CompileError::Internal(
-                        "Expecting associated type",
-                        item_ref.span(engines),
-                    )));
-                }
-            }
-            TypeInfo::Ref {
-                referenced_type: mut ty,
-                to_mutable_value,
-            } => {
-                ty.type_id = self
-                    .resolve(
-                        handler,
-                        engines,
-                        namespace,
-                        mod_path,
-                        ty.type_id,
-                        span,
-                        enforce_type_arguments,
-                        None,
-                        self_type,
-                    )
-                    .unwrap_or_else(|err| {
-                        engines
-                            .te()
-                            .insert(engines, TypeInfo::ErrorRecovery(err), None)
-                    });
-
-                engines.te().insert(
-                    engines,
-                    TypeInfo::Ref {
-                        to_mutable_value,
-                        referenced_type: ty.clone(),
-                    },
-                    None,
-                )
-            }
-            _ => type_id,
-        };
-
-        let mut type_id = type_id;
-        type_id.subst(
-            &self.type_subst(),
-            &SubstTypesContext::new(engines, !self.collecting_unifications()),
-        );
-
-        Ok(type_id)
-    }
-
-    fn resolve_qualified_call_path(
-        &self,
-        handler: &Handler,
-        engines: &Engines,
-        namespace: &Namespace,
-        mod_path: &ModulePath,
-        qualified_call_path: &QualifiedCallPath,
-    ) -> Result<ty::TyDecl, ErrorEmitted> {
-        let type_engine = engines.te();
-        if let Some(qualified_path_root) = qualified_call_path.clone().qualified_path_root {
-            let root_type_id = match &&*type_engine.get(qualified_path_root.ty.type_id) {
-                TypeInfo::Custom {
-                    qualified_call_path,
-                    type_arguments,
-                    ..
-                } => {
-                    let type_decl = self.resolve_call_path(
-                        handler,
-                        engines,
-                        namespace,
-                        mod_path,
-                        &qualified_call_path.clone().to_call_path(handler)?,
-                    )?;
-                    type_decl_opt_to_type_id(
-                        handler,
-                        self.engines(),
-                        self.namespace(),
-                        self,
-                        Some(type_decl),
-                        &qualified_call_path.call_path,
-                        &qualified_path_root.ty.span(),
-                        EnforceTypeArguments::No,
-                        mod_path,
-                        type_arguments.clone(),
-                        self.self_type(),
-                    )?
-                }
-                _ => qualified_path_root.ty.type_id,
-            };
-
-            let as_trait_opt = match &&*type_engine.get(qualified_path_root.as_trait) {
-                TypeInfo::Custom {
-                    qualified_call_path: call_path,
-                    ..
-                } => Some(
-                    call_path
-                        .clone()
-                        .to_call_path(handler)?
-                        .to_fullpath(self.engines(), self.namespace()),
-                ),
-                _ => None,
-            };
-
-            namespace
-                .root
-                .resolve_call_path_and_root_type_id(
-                    handler,
-                    engines,
-                    &namespace.root.module,
-                    root_type_id,
-                    as_trait_opt,
-                    &qualified_call_path.call_path,
-                    self.self_type(),
-                )
-                .map(|decl| decl.expect_typed())
-        } else {
-            self.resolve_call_path(
-                handler,
-                engines,
-                namespace,
-                mod_path,
-                &qualified_call_path.call_path,
-            )
-        }
-    }
-
-    /// Resolve a symbol that is potentially prefixed with some path, e.g. `foo::bar::symbol`.
-    ///
-    /// This will concatenate the `mod_path` with the `call_path`'s prefixes and
-    /// then calling `resolve_symbol` with the resulting path and call_path's suffix.
-    ///
-    /// The `mod_path` is significant here as we assume the resolution is done within the
-    /// context of the module pointed to by `mod_path` and will only check the call path prefixes
-    /// and the symbol's own visibility.
-    fn resolve_call_path(
-        &self,
-        handler: &Handler,
-        engines: &Engines,
-        namespace: &Namespace,
-        mod_path: &ModulePath,
-        call_path: &CallPath,
-    ) -> Result<ty::TyDecl, ErrorEmitted> {
-        let (decl, mod_path) = namespace.root.resolve_call_path_and_mod_path(
-            handler,
-            self.engines,
-            mod_path,
-            call_path,
-            self.self_type,
-        )?;
-        let decl = decl.expect_typed();
-
-        // In case there is no mod path we don't need to check visibility
-        if mod_path.is_empty() {
-            return Ok(decl);
-        }
-
-        // In case there are no prefixes we don't need to check visibility
-        if call_path.prefixes.is_empty() {
-            return Ok(decl);
-        }
-
-        // check the visibility of the call path elements
-        // we don't check the first prefix because direct children are always accessible
-        for prefix in iter_prefixes(&call_path.prefixes).skip(1) {
-            let module = namespace.lookup_submodule_from_absolute_path(handler, engines, prefix)?;
-            if module.visibility().is_private() {
-                let prefix_last = prefix[prefix.len() - 1].clone();
-                handler.emit_err(CompileError::ImportPrivateModule {
-                    span: prefix_last.span(),
-                    name: prefix_last,
-                });
-            }
-        }
-
-        // check the visibility of the symbol itself
-        if !decl.visibility(self.engines.de()).is_public() {
-            handler.emit_err(CompileError::ImportPrivateSymbol {
-                name: call_path.suffix.clone(),
-                span: call_path.suffix.span(),
-            });
-        }
-
-        Ok(decl)
     }
 }
